Updated environments organization
