+++ fitr/environments/environments.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+import re
+import numpy as np
+import networkx as nx
+import matplotlib.pyplot as plt
+from scipy.stats import multivariate_normal
+from scipy.linalg import circulant
+
+
+
+
+#===============================================================================
+#   SIMPLE FUNCTIONS TO RUN ENVIRONMENTS
+#===============================================================================
